requirements.txt
black
requests
pylint
python-dotenv